Improve API, add mutex
The callback signature was changed to pass MQTTPublishInfo_t,
which contains the previously passed data.
Added a mutex to protect the API against concurrent access.
Added a semaphore to check for publish QoS.
Added missing QoS parameter to unsubscribe.
It is now possible to call processResponses with a single loop.